+++ src/page/LandingPage/components/TimeTableSection.tsx
@@ -1,0 +1,33 @@
+import section101 from '../../../../public/landing/section1-1.png';
+import section102 from '../../../../public/landing/section1-2.png';
+
+export default function TimeTableSection() {
+  return (
+    <section id="section1" className="flex flex-col gap-24">
+      <div>
+        <div className="relative inline-block text-[1.5vw] font-semibold">
+          <span className="relative z-10">시간표 설정화면</span>
+          <span className="absolute bottom-0 left-0 z-0 h-4 w-full bg-brand-main/70"></span>
+        </div>
+        <h1 className="mt-4 text-left text-[2.75vw] font-bold">
+          드래그 앤 드롭으로
+          <br /> 간편하게 시간표 구성
+        </h1>
+      </div>
+
+      <img src={section101} alt="section101" className="w-full" />
+      <img src={section102} alt="section102" className="mt-5 w-full" />
+      <div className="flex flex-col items-center gap-1">
+        <p className="mb-2 text-center text-[1.75vw] font-semibold">
+          두가지 타이머
+        </p>
+        <p className="text-center text-[1.5vw] font-medium text-neutral-600">
+          일반형과 자유토론형 타이머로,
+        </p>
+        <p className="text-center text-[1.5vw] font-medium text-neutral-600">
+          다양한 토론 방식을 지원해요.
+        </p>
+      </div>
+    </section>
+  );
+}
